Move all workload utilities to their own dir
